requirements.txt
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Django==5.1.11
django-tailwind==3.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
entrypoints==0.4
Faker==37.4.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pillow==11.2.1
protobuf==3.20.3
psycopg2-binary==2.9.10
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.9.0
Pympler==1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.22.0
tenacity==8.5.0
toml==0.10.2
toolz==1.0.0
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
zipp==3.23.0
